socioxplorer-backend/requirements.txt
astor==0.8.1
emoji==2.10.1
fasttext-langdetect==1.0.5
Flask-SQLAlchemy==3.0.5
Flask==2.3.2
gunicorn==21.2.0
h5py==3.11.0
html2text
ipython==8.12.3
matplotlib-inline==0.1.7
nltk==3.8.1
numba==0.58.1
numpy==1.24.3
pandas==2.0.3
parso==0.8.4 
prompt-toolkit==3.0.47
pure-eval==0.2.3
pygments==2.18.0
pysolr==3.9.0
python-dotenv==1.0.1
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard-data-server==0.7.2
tensorboard==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow==2.13.1
traitlets==5.14.3
transformers==4.42.4
umap-learn==0.5.3
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
psutil==6.1.0
schedule==1.2.2
tensorflow-probability==0.21.0
